Use context to safely break out of loops and cancel tickers
Wherever possible, use context to ensure that the code path cannot enter
a forever loop as a result of a developer error. This would have the
added benefit in some cases of avoiding an additional iteration when
context is done.

Add defer statements to stop tickers as required.

Fix a minor error type check.