Add ramping loadtests to test cluster  (#15)
This adds a ramping loadtest that can be used test progressive levels of
load on a loadtest cluster to identify the point at which we experience
failure to meet SLO (which is p90 block times of 1s in this case). This
can be run with any scenario as specified, and we can configure later CI
to run loadtests for varying scenarios to evaluate TPS under different
load profiles.